move filtered list management to TodoList component, change filter input to dropdown menu
