+++ src/components/Todo/Todo.jsx
@@ -12,55 +12,22 @@ export default function Todo(){
     console.log('Todo.jsx rendered')
 
     const [todos, dispatch] = useLocalStorageReducer('todos',todoReducer,[{title:"My First Task!", id:-1, isDone:false}]);
-    const [filtered, dispatchFilter] = useReducer(filterReducer,[])
-    const [filterType, setFilterType] = useState('all')
+    // const [filtered, dispatchFilter] = useReducer(filterReducer,[])
+    // const [filterType, setFilterType] = useState('all')
 
-    
-    useEffect(()=>{
-        dispatchFilter({type: filterType, payload:todos})
-    },[todos, filterType])
+
 
     useEffect(()=>{
         localStorage.setItem('todos',JSON.stringify(todos))
     },[todos])
 
-    const showPending = () =>{
-        setFilterType('pending')
-    }
-
-    const showAll = () =>{
-        setFilterType('all')
-    }
-
     return(
         <div id="todo">
             <TodoContext.Provider value={dispatch}>
                 <AddTodo />
-                <div>
-                    <Button onClickFunc={showAll}>
-                        {`Show All (${todos.length})`}
-                    </Button>
-                    <Button onClickFunc={showPending} >
-                        {`Pending (${todos.filter((todo)=>todo.isDone===false).length})`}
-                    </Button>
-                </div>
-                <TodoList list={filtered} /> 
+                <TodoList todos={todos} /> 
             </TodoContext.Provider> 
         </div>
     );
 }
 
-function filterReducer(state,action){
-    const {type, payload} = action;
-    switch(type){
-        case 'pending':{
-            return [...payload.filter((todo)=>todo.isDone===false)];
-        }
-        case 'all':{
-            return [...payload];
-        }
-        default:{
-            break;
-        }
-    }
-}
