+++ src/components/Todo/TodoList/TodoList.jsx
@@ -1,10 +1,47 @@
+import { useReducer, useState, useEffect } from "react";
 import TodoItem from "../TodoItem/TodoItem";
+import Button from "../../common/Button";
 import "./TodoList.scss"
 
-export default function TodoList({list}){
+export default function TodoList({todos}){
+
+    const [filtered, dispatchFilter] = useReducer(filterReducer,[])
+    const [filterType, setFilterType] = useState('all')
+  
+    useEffect(()=>{
+        dispatchFilter({type: filterType, payload:todos})
+    },[todos, filterType])
+
+    const showPending = () =>{
+        setFilterType('pending')
+    }
+
+    const showAll = () =>{
+        setFilterType('all')
+    }
+
+    const handleDropdownChange = (e)=>{
+        setFilterType(e.target.value)
+    }
+
     return(
-        <ul className="todolist-container">
-            {list.map((todo) => (
+        <>
+            <div className="list-filters">
+                {/* <Button onClickFunc={showAll}>
+                    {`Show All (${todos.length})`}
+                </Button>
+                <Button onClickFunc={showPending} >
+                    {`Pending (${todos.filter((todo)=>todo.isDone===false).length})`}
+                </Button> */}
+                <label for="status">Status</label>
+                <select id="status" onChange={handleDropdownChange}>
+                    <option value="all">All</option>
+                    <option value="pending">Pending</option>
+                    <option value="done">Done</option>
+                </select>
+            </div>
+            <ul className="todolist-container">
+            {filtered.map((todo) => (
                 <li 
                 key={todo.id}>
                     <TodoItem 
@@ -14,6 +51,26 @@ export default function TodoList({list}){
                 </li>
                 )
             )}
-        </ul> 
+            </ul> 
+        </>
+        
     );
+}
+
+function filterReducer(state,action){
+    const {type, payload} = action;
+    switch(type){
+        case 'pending':{
+            return [...payload.filter((todo)=>todo.isDone===false)];
+        }
+        case 'done':{
+            return [...payload.filter((todo)=>todo.isDone===true)];
+        }
+        case 'all':{
+            return [...payload];
+        }
+        default:{
+            break;
+        }
+    }
 }
